Improve the matlab script for plotting
Improve the looking of the plots (title, names
of axis, ...). Add a section to read variables
of interest from files and compute epsilon_r.
Add a section creating a results directory where
plots and inputs of the experiment can be saved.
Add some lines of code to save automatically the
plots in the results folder created above. Give
the possibility to choose times to plot. Comment
away some sections of the script that don't work.